Extract navigation tools (hover, completions, declaration_file, references)
Tool modules import non-patched utilities directly from their source
modules; server-defined helpers and test-patched symbols are referenced via
the server module so monkeypatching is unaffected.
